Made the .npmigore file more inclusive
I think including the "/migrations" folder might not be a bad idea,
given that OpenZeppelin included it in their "openzeppelin-solidity"
library. I'm kind of just following their example at this point, so it
may not actually be useful in any way, but I figure there doesn't seem
to be any reason to leave it out.