Added the "Killable" management archetype
"Killable" allows its children to be removed from the blockchain once
their owner decides to kill them. They also support forwarding any
remaining contract funds to a particular address once the contract has
been killed.

In the process of creating and testing "Killable" I also had to make a
number of modifications to the package itself, as well as to the
migration system.

Included also is the appropriate documentation for "Killable" along with
some of the framework I created in the process of setting up the
experimental testing pages which should be coming soon.