Added references within the README to the unstable branch
Users should be aware that the project has unstable content publicly
available. Ideally, they shouldn't be using any of it, but they should
know it's there to help build support for it and to give users the
opportunity to develop their applications in anticipation of the new
modules which are under development.