Switched the order of the "Destroyed" event
After some testing, I have determined that attempting to fire an event
after calling "selfdestruct" will not reliably cause that event to be
fired. Makes sense. I have corrected for the issue by firing the event
just before the selfdestruct call is made.

This could potentially cause problems further down the road if the event
persists in the event that the contract runs out of gas prior to the
actual destruction of the contract. I don't think that would happen,
given the nature of the out-of-gas exception, but I could be wrong. This
wil require further testing.