Changed the testing pages to rely on the "contracts/examples" directory
A lot of the contracts that I plan to add to the project in the
relatively near future are very abstract, and as such, were not
previously testable using the testing pages. I have decided to remedy
the situation by adding the "contracts/examples" directory, which will
contain sample implementations of each contract. The testing pages will
then link directly to these example contracts, rather than to the parent
contracts themselves.

As such, it would be a good idea to test the parent contracts and
libraries directly using a unit testing suite, given that the testing
pages will likely no longer be sufficient.